More work on AddBoxDialogFragment to implement more user input. Users can now choose if the box will be stored in fridge or freezer as well as select ingredients from the list. Still work to do on the ingredient selection, so that it is visible which ingredients has been chosen.

+++ app/src/main/java/com/fannyarvid/foodorganizer/MainActivity.java
@@ -94,7 +94,7 @@ public class MainActivity extends ActionBarActivity
 
     @Override
     public void onAddBoxDialogPositiveClick(
-            String boxName, int julianDay, int storageType, int hasBeenInFreezer, int[] ingredientsId) {
+            String boxName, int julianDay, int storageType, int hasBeenInFreezer, long[] ingredientsId) {
         ContentValues boxValues = new ContentValues();
         boxValues.put(FoodContract.BoxEntry.COLUMN_BOX_NAME, boxName);
         boxValues.put(FoodContract.BoxEntry.COLUMN_DATE, julianDay);
